NOAA transitioned to a class. Decomposing functions into individual source files.
