+++ include/mov_flags.h
@@ -1,0 +1,7 @@
+#ifndef MOV_FLAGS_H
+#define MOV_FLAGS_H
+
+#define MOV_NULL_TS -8880
+
+#endif // MOV_FLAGS_H
+
