+++ include/noaa.h
@@ -2,13 +2,13 @@
 #define NOAA_H
 
 #include <MetOceanViewer.h>
+#include <mov_errors.h>
+#include <mov_flags.h>
 #include <QtWebKitWidgets>
 #include <QNetworkInterface>
 #include <QUrl>
 #include <QtNetwork>
 #include <QVector>
-
-#define NOAA_ERR_INVALIDDATERANGE -9990
 
 class noaa : public QMainWindow
 {
@@ -17,20 +17,43 @@ public:
     noaa();
 
     //...Public Functions
-    int fetchNOAAData();
+    int fetchNOAAData(QVector<QString> &javascript);
+    int PlotNOAAResponse(QVector<QString> &javascript);
+
+    //...Structures
+    struct NOAAStationData
+    {
+        QDate Date;
+        QTime Time;
+        double value;
+    };
+
+    //...Public Variables
+    int NOAAMarkerID;
+    int ProductIndex;
+
+    double CurrentNOAALat;
+    double CurrentNOAALon;
+
+    QString CurrentNOAAStationName;
+    QString Datum;
+    QString Units;
+
+    QDateTime StartDate;
+    QDateTime EndDate;
+
+    QVector< QVector<NOAAStationData> > CurrentNOAAStation;
 
 private:
 
     //...Private Functions
+    QString FormatNOAAResponse(QVector<QByteArray> Input, QString &ErrorString, int index);
     void ReadNOAAResponse(QNetworkReply *reply, int index, int index2);
-    void PlotNOAAResponse();
+    int retrieveProduct(int type, QString &Product, QString &Product2);
 
     //...Private Variables
     QVector< QVector<QByteArray> > NOAAWebData;
-    int NOAAMarkerID,ProductIndex;
-    double CurrentNOAALat,CurrentNOAALon;
-    QString CurrentNOAAStationName,Datum,Units;
-    QDateTime StartDate,EndDate;
+
 };
 
 #endif // NOAA_H
